+++ Masal Zamanı/GADBannerTableViewController.m
@@ -73,7 +73,7 @@
 {
     static NSString *CellIdentifier = @"Cell";
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:CellIdentifier];
-    int rowIndex = [indexPath row];
+    long rowIndex = [indexPath row];
     
     if (cell == nil) {
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:CellIdentifier];
